requirements.txt
arrow==1.3.0
aiofiles==23.2.1
dependency-injector==4.41.0
uvicorn==0.29.0
fastapi==0.110.0
fastapi-sso==0.13.0
google-auth==2.29.0
pydantic[email]==2.6.4
bcrypt==4.1.2
celery==5.3.6
redis==5.0.3
pillow==10.3.0
python-dotenv==1.0.1
python-jose==3.3.0
ruff==0.3.4
motor==3.3.2
minio==7.2.5
icecream==2.1.3
python-multipart==0.0.9
odmantic==1.0.1
pydantic-settings==2.2.1